Combined region handling bug fixes...
Sometimes the xml_pre_child_layering_index would be nil/Undetermined for a child region, and thus not sortable. --> Fallback to 0 now.

Also, parent regions were still being added to and checked for missing_sources, when of course parents we're reconstructing have no TSV2 match (with IDs) in and of themselves.